Move container files to pages folder
