--- app/containers/App.js
@@ -1,14 +1,0 @@
-// @flow
-import * as React from 'react';
-
-type Props = {
-  children: React.Node
-};
-
-export default class App extends React.Component<Props> {
-  props: Props;
-
-  render() {
-    return <div>{this.props.children}</div>;
-  }
-}
