+++ app/pages/dashboard/index.jsx
@@ -1,9 +1,9 @@
 // @flow
 import React, { Component } from 'react';
-import Dashboard from '../components/dashboard/Dashboard';
-import SiteMenuBar from '../components/SiteMenuBar';
-import SiteNavbar from '../components/SiteNavbar';
-import ErrorBoundary from '../components/ErrorBoundary';
+import Dashboard from '../../components/dashboard/Dashboard';
+import SiteMenuBar from '../../components/SiteMenuBar';
+import SiteNavbar from '../../components/SiteNavbar';
+import ErrorBoundary from '../../components/ErrorBoundary';
 
 export default class HomePage extends Component {
   componentDidMount() {
