+++ app/pages/login/index.jsx
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component } from 'react';
-import Login from '../components/Login';
+import Login from '../../components/Login';
 
 export default class LoginPage extends Component {
   componentDidMount() {
