+++ app/pages/series/index.js
@@ -1,0 +1,2 @@
+export { default as SeriesPage } from './series';
+export { default as SeriesListPage } from './seriesList';
